backend/requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
django-import-export==3.3.8
django-leaflet==0.29.1
djangorestframework==3.15.0
postgis==1.0.4
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
daphne==4.0.0
django-rest-passwordreset==1.4.1











